# papahana_flask_server_demo/papahana/models/program.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from papahana.models.base_model_ import Model
from papahana import util


class Program(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, name: str=None, sem_id: str=None, containers: List[str]=None):  # noqa: E501
        """Program - a model defined in Swagger

        :param name: The name of this Program.  # noqa: E501
        :type name: str
        :param sem_id: The sem_id of this Program.  # noqa: E501
        :type sem_id: str
        :param containers: The containers of this Program.  # noqa: E501
        :type containers: List[str]
        """
        self.swagger_types = {
            'name': str,
            'sem_id': str,
            'containers': List[str]
        }

        self.attribute_map = {
            'name': 'name',
            'sem_id': 'sem_id',
            'containers': 'containers'
        }
        self._name = name
        self._sem_id = sem_id
        self._containers = containers

    @classmethod
    def from_dict(cls, dikt) -> 'Program':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Program of this Program.  # noqa: E501
        :rtype: Program
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this Program.


        :return: The name of this Program.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Program.


        :param name: The name of this Program.
        :type name: str
        """

        self._name = name

    @property
    def sem_id(self) -> str:
        """Gets the sem_id of this Program.


        :return: The sem_id of this Program.
        :rtype: str
        """
        return self._sem_id

    @sem_id.setter
    def sem_id(self, sem_id: str):
        """Sets the sem_id of this Program.


        :param sem_id: The sem_id of this Program.
        :type sem_id: str
        """

        self._sem_id = sem_id

    @property
    def containers(self) -> List[str]:
        """Gets the containers of this Program.


        :return: The containers of this Program.
        :rtype: List[str]
        """
        return self._containers

    @containers.setter
    def containers(self, containers: List[str]):
        """Sets the containers of this Program.


        :param containers: The containers of this Program.
        :type containers: List[str]
        """

        self._containers = containers